feat: PageResponse 클래스를 Common으로 위치 변경, 정적 팩토리 메서드로 수정
