+++ Main/src/main/java/com/seveneleven/board/controller/BoardController.java
@@ -1,9 +1,10 @@
 package com.seveneleven.board.controller;
 
 import com.seveneleven.board.dto.*;
-import com.seveneleven.board.service.PostService;
+import com.seveneleven.board.service.PostServiceImpl;
 import com.seveneleven.entity.board.constant.PostFilter;
 import com.seveneleven.response.APIResponse;
+import com.seveneleven.response.PageResponse;
 import com.seveneleven.response.SuccessCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.MediaType;
@@ -11,7 +12,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.awt.print.Pageable;
 import java.util.List;
 
 @RestController
@@ -19,7 +19,7 @@ import java.util.List;
 @RequestMapping("/api/posts")
 public class BoardController implements BoardDocs {
 
-    private final PostService postService;
+    private final PostServiceImpl postService;
 
     /**
      * 함수명 : selectList()
@@ -27,10 +27,10 @@ public class BoardController implements BoardDocs {
      */
     @GetMapping("/steps/{projectStepId}")
     public ResponseEntity<APIResponse<PageResponse<PostListResponse>>> selectList (@PathVariable Long projectStepId,
-                                                             @RequestParam(defaultValue = "0") Integer page,
-                                                             @RequestParam(required = false) String keyword,
-                                                             @RequestParam(required = false) PostFilter filter
-                                                             // todo: 정렬기준 추후 추가 예정
+                                                                                   @RequestParam(defaultValue = "0") Integer page,
+                                                                                   @RequestParam(required = false) String keyword,
+                                                                                   @RequestParam(required = false) PostFilter filter
+                                                                                   // todo: 정렬기준 추후 추가 예정
     ) throws Exception {
         PageResponse<PostListResponse> postList = postService.selectList(projectStepId, page, keyword, filter);
